feat: update PointMeta, FilterMeta, and ProjectMeta models for improved JSON serialization; rename fields for clarity and consistency

+++ lib/db/src/meta/point_meta.dart
@@ -6,8 +6,8 @@ part 'point_meta.g.dart'; // JSON support
 @freezed
 class PointMeta with _$PointMeta {
   const factory PointMeta({
-    required int x,
-    required int y,
+    required double x,
+    required double y,
   }) = _PointMeta;
 
   factory PointMeta.fromJson(Map<String, Object?> json) => _$PointMetaFromJson(json);
